requirements.txt
flask==2.3.3
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
matplotlib==3.7.2
lime==0.2.0.1
scikit-image==0.21.0
pillow==10.0.0
werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
